Changes in organization, functionality, and tests
