--- egypt-time-weather/requirements.txt
@@ -1,4 +1,0 @@
-Flask==2.0.1
-pytz==2021.1
-requests==2.25.1
-pytest==6.2.4
