+++ requirements.txt
@@ -1,0 +1,18 @@
+blinker==1.8.2
+certifi==2024.2.2
+charset-normalizer==3.3.2
+click==8.1.7
+Flask==3.0.3
+idna==3.7
+iniconfig==2.0.0
+itsdangerous==2.2.0
+Jinja2==3.1.4
+MarkupSafe==2.1.5
+packaging==24.0
+pluggy==1.5.0
+pytest==8.2.1
+python-dotenv==1.0.1
+pytz==2024.1
+requests==2.32.2
+urllib3==2.2.1
+Werkzeug==3.0.3
