+++ tests/test_app.py
@@ -1,0 +1,19 @@
+import json
+
+import pytest
+
+from app import app
+
+
+@pytest.fixture
+def client():
+    with app.test_client() as client:
+        yield client
+
+
+def test_home(client):
+    """Test the home route"""
+    response = client.get('/')
+
+    assert response.status_code == 200
+    assert 'Weather App' in response.data.decode()
